// Project_Apex/frontend/src/api/models/PendingWithdrawal.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PendingWithdrawal = {
    id: string;
    user_id: string;
    email: string;
    amount: number;
    description: string;
    created_at: string;
    status: string;
    source: string;
    investment_id?: string | null;
    plan_name?: string | null;
};